Refactor for clarity and return error
Found that an exception would occur if my list was empty and I tried
to view a tail node that was nil. I added a simple return message to
indicate what happened. I changed the "remove_at" method to be a
little more clear on what the return value was supposed to be if the
head node was removed.